requirements.txt
# Core Framework
streamlit==1.39.0
python-dotenv==1.0.1

# Google AI
google-generativeai==0.8.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
Pillow==10.4.0
pytesseract==0.3.13
PyMuPDF==1.24.13

# Vector Database & Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# LangChain
langchain==0.3.13
langchain-google-genai==2.0.5
langchain-community==0.3.13
langchain-chroma==0.1.4

# PDF Generation
reportlab==4.2.5
markdown==3.7

# Data Processing
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5

# API & Utils
requests==2.32.3
aiofiles==24.1.0

# Monitoring (Optional)
loguru==0.7.3
